requirements.txt
streamlit>=1.30.0
requests>=2.31.0
openai>=1.10.0
pydantic>=2.5.0
python-dotenv>=1.0.0
bs4
